+++ ui/src/components/navigation/desktop/UserProfileDropdown.tsx
@@ -29,7 +29,7 @@ import { useAuthStore } from "@/common/state/features/auth/auth.slice";
 const UserProfileDropdownWrapper: FC<{ children: ReactNode }> = ({ children }) => {
   const { user, logout } = useAuth0();
   const navigate = useNavigate();
-  const { isAdmin } = useAuthStore();
+  const { isAdmin, isGuest } = useAuthStore();
 
   return (
     <DropdownMenu>
@@ -52,10 +52,12 @@ const UserProfileDropdownWrapper: FC<{ children: ReactNode }> = ({ children }) =
             <ShoppingCart className="w-4 h-4 mr-2" />
             <span>My Cart</span>
           </DropdownMenuItem>
-          <DropdownMenuItem onClick={() => navigate("/my-profile")}>
-            <User className="w-4 h-4 mr-2" />
-            <span>My Orders</span>
-          </DropdownMenuItem>
+          {!isGuest && (
+            <DropdownMenuItem onClick={() => navigate("/my-profile")}>
+              <User className="w-4 h-4 mr-2" />
+              <span>My Orders</span>
+            </DropdownMenuItem>
+          )}
         </DropdownMenuGroup>
         {isAdmin && (
           <Fragment>